Refactor code to use module and separate responsibilities
